Fix: Resolve blank page issue by unifying Document type
This commit addresses a critical bug that caused a blank white page to be displayed on the production site. The root cause was a `TypeError` on the client-side, triggered by an inconsistency between the data structure expected by the frontend and the one provided by the backend API.

Key changes:
- **Unified `Document` Type:** Consolidated the `Document` type definition into a single source of truth in `shared/api.ts` to eliminate conflicts between the client, server, and Netlify functions.
- **Corrected Data Mapping:** Updated the Netlify function (`netlify/functions/google-sheets.ts`) to use the shared `Document` type and correctly map the `perihal` field (previously `subject`), resolving the direct cause of the client-side crash.
- **Improved Client-Side Resilience:** Added a defensive check in the client's data store (`client/lib/documentStore.ts`) to prevent crashes from malformed `expeditionHistory` data.
- **Fixed Type Errors:** Resolved numerous TypeScript errors across the codebase that resulted from the type consolidation, including incorrect imports, missing properties on created objects, and unsafe method calls on union types. This improves the overall stability and maintainability of the code.

All type checks and tests now pass, confirming the validity of the fix and ensuring no regressions were introduced.
