+++ client/pages/DocumentList.tsx
@@ -2,13 +2,13 @@ import React, { useState, useEffect } from "react";
 import {
   Search,
   FileText,
-  User,
   RefreshCw,
   AlertCircle,
   Database,
   Calendar,
 } from "lucide-react";
-import { useDocumentStore, Document } from "../lib/documentStore";
+import { useDocumentStore } from "../lib/documentStore";
+import { Document } from "../lib/types";
 import { format } from "date-fns";
 import { id } from "date-fns/locale";
 import { cn } from "../lib/utils";
